src/GreenQRCode/requirements.txt
arrow==1.1.1
asteval==0.9.25
astor==0.8.1
astroid==2.5.6
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
boto3==1.18.11
botocore==1.21.11
cachetools==4.2.2
Cerberus==1.3.4
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
DateTime==4.2
decorator==4.4.2
distlib==0.3.2
docopt==0.6.2
easyocr==1.4
et-xmlfile==1.1.0
fake-useragent==0.1.11
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
frontmatter==3.0.7
future==0.18.2
gast==0.3.3
gevent==21.1.2
greenlet==1.1.0
identify==2.2.13
idna==2.10
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.6.4
iso8601==0.1.14
isort==5.9.1
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
leancloud==2.9.4
lmdb==1.2.1
lmfit==1.0.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
mccabe==0.6.1
monty==2021.7.8
msgpack-python==0.5.6
networkx==2.5.1
nodeenv==1.6.0
numexpr==2.7.3
numpy==1.19.3
opencv-contrib-python==4.4.0.46
opencv-python==4.5.3.56
openpyxl==3.0.7
paddleocr==2.2.0.2
paddlepaddle==2.0.0
pandas==1.3.1
Pillow==8.2.0
platformdirs==2.2.0
powerline-status==2.7
pre-commit==2.14.0
premailer==3.10.0
protobuf==3.17.3
pyclipper==1.3.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydash==5.0.2
pyflakes==2.3.1
pylint==2.8.3
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-frontmatter==1.0.0
python-Levenshtein==0.12.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.1
pyzbar==0.1.8
qiniu==7.3.1
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
secure-cookie==0.1.0
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.16.0
tables==3.6.1
threadpoolctl==2.2.0
tifffile==2021.8.8
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
uncertainties==3.1.6
urllib3==1.26.5
virtualenv==20.7.2
visualdl==2.2.0
watchtower==1.0.6
Werkzeug==2.0.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
